requirement.txt
asgiref==3.6.0
autopep8==2.0.2
certifi==2023.5.7
charset-normalizer==3.1.0
crispy-bootstrap4==2022.1
distlib==0.3.6
Django==4.2.1
django-crispy-forms==2.0
django-extensions==3.2.3
filelock==3.12.2
idna==3.4
Pillow==9.5.0
platformdirs==3.8.0
pycodestyle==2.10.0
pydotplus==2.0.2
pyparsing==3.1.0
requests==2.31.0
sqlparse==0.4.4
sweetify==2.3.1
tzdata==2023.3
urllib3==2.0.3
virtualenv==20.23.1
